Remove reth-based types and move more types into trin-types.
