+++ ethportal-api/src/types/mod.rs
@@ -1,5 +1,3 @@
-pub mod content_item;
-pub mod content_key;
 pub mod discv5;
 pub mod enr;
 pub mod portal;
