+++ trin-core/src/portalnet/storage.rs
@@ -19,7 +19,7 @@ use rusqlite::params;
 use tracing::{debug, info};
 
 use crate::{portalnet::types::messages::ProtocolId, utils::db::get_data_dir};
-use ethportal_api::types::content_key::{HistoryContentKey, OverlayContentKey};
+use trin_types::content_key::{HistoryContentKey, OverlayContentKey};
 use trin_types::distance::{Distance, Metric, XorMetric};
 use trin_utils::bytes::{hex_decode, hex_encode};
