+++ trin-types/Cargo.toml
@@ -19,7 +19,6 @@ discv5 = { version = "0.2.1", features = ["serde"]}
 eth_trie = "0.1.0"
 ethereum-types = "0.12.1"
 ethnum = "1.3.2"
-ethportal-api = { path = "../ethportal-api" }
 eth2_ssz = "0.4.0"
 eth2_ssz_derive = "0.3.0"
 eth2_ssz_types = "0.2.1"
@@ -30,10 +29,13 @@ rlp-derive = "0.1.0"
 serde = { version = "1.0.150", features = ["derive"] }
 serde_json = "1.0.89"
 serde_yaml = "0.9.17"
+sha2 = "0.10.1"
 sha3 = "0.9.1"
 snap = "1.1.0"
 structopt = "0.3.26"
-tree_hash = "0.5.0"
+#tree_hash = "0.5.0"
+tree_hash = "0.4.0"
+tree_hash_derive = "0.4.0"
 tokio = { version = "1.14.0", features = ["full"] }
 trin-utils = { path = "../trin-utils" }
 ureq = { version = "2.5.0", features = ["json"] }
