+++ trin-types/src/execution/accumulator.rs
@@ -1,0 +1,19 @@
+use ethereum_types::U256;
+use serde::{Deserialize, Serialize};
+use ssz_derive::{Decode, Encode};
+use ssz_types::{typenum, VariableList};
+use tree_hash_derive::TreeHash;
+
+/// SSZ List[HeaderRecord, max_length = EPOCH_SIZE]
+/// List of (block_number, block_hash) for each header in the current epoch.
+pub type EpochAccumulator = VariableList<HeaderRecord, typenum::U8192>;
+
+/// Individual record for a historical header.
+/// Block hash and total difficulty are used to validate whether
+/// a header is canonical or not.
+/// Every HeaderRecord is 64bytes.
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Decode, Encode, Deserialize, Serialize, TreeHash)]
+pub struct HeaderRecord {
+    pub block_hash: tree_hash::Hash256,
+    pub total_difficulty: U256,
+}
