+++ trin-types/src/execution/mod.rs
@@ -1,3 +1,4 @@
+pub mod accumulator;
 pub mod block_body;
 pub mod header;
 pub mod receipts;
