+++ trin-types/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bytes;
 pub mod cli;
 pub mod consensus;
+pub mod content_item;
 pub mod content_key;
 pub mod distance;
 pub mod enr;
